+++ packages/cli/CHANGELOG.md
@@ -1,13 +1,26 @@
 # @auto-engineer/cli
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @auto-engineer/narrative@0.15.1
-  - @auto-engineer/file-store@0.15.1
-  - @auto-engineer/pipeline@0.15.1
+  - @auto-engineer/file-store@0.17.0
+  - @auto-engineer/narrative@0.17.0
+  - @auto-engineer/pipeline@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/narrative@0.16.0
+  - @auto-engineer/pipeline@0.16.0
+  - @auto-engineer/file-store@0.16.0
 
 ## 0.15.0