+++ packages/create-auto-app/CHANGELOG.md
@@ -1,11 +1,22 @@
 # create-auto-app
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @auto-engineer/id@0.15.1
+  - @auto-engineer/id@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/id@0.16.0
 
 ## 0.15.0