+++ packages/information-architect/CHANGELOG.md
@@ -1,13 +1,26 @@
 # @auto-engineer/information-architect
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @auto-engineer/narrative@0.15.1
-  - @auto-engineer/ai-gateway@0.15.1
-  - @auto-engineer/message-bus@0.15.1
+  - @auto-engineer/message-bus@0.17.0
+  - @auto-engineer/ai-gateway@0.17.0
+  - @auto-engineer/narrative@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/narrative@0.16.0
+  - @auto-engineer/ai-gateway@0.16.0
+  - @auto-engineer/message-bus@0.16.0
 
 ## 0.15.0