+++ packages/message-bus/CHANGELOG.md
@@ -1,6 +1,12 @@
 # @auto-engineer/message-bus
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
+
+## 0.16.0
 
 ## 0.15.0