+++ packages/narrative/CHANGELOG.md
@@ -1,15 +1,31 @@
 # @auto-engineer/flow
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
+- add first-class Module support for type ownership
 
 ### Patch Changes
 
-- add first-class Module support for type ownership
+- Updated dependencies []:
+  - @auto-engineer/message-bus@0.17.0
+  - @auto-engineer/file-store@0.17.0
+  - @auto-engineer/id@0.17.0
+
+## 0.16.0
+
+### Minor Changes
+
+- Adds new endpoints for on.auto
+
+### Patch Changes
 
 - Updated dependencies []:
-  - @auto-engineer/file-store@0.15.1
-  - @auto-engineer/id@0.15.1
-  - @auto-engineer/message-bus@0.15.1
+  - @auto-engineer/file-store@0.16.0
+  - @auto-engineer/id@0.16.0
+  - @auto-engineer/message-bus@0.16.0
 
 ## 0.15.0