+++ packages/pipeline/CHANGELOG.md
@@ -1,12 +1,28 @@
 # @auto-engineer/pipeline
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @auto-engineer/file-store@0.15.1
-  - @auto-engineer/message-bus@0.15.1
+  - @auto-engineer/message-bus@0.17.0
+  - @auto-engineer/file-store@0.17.0
+
+## 0.16.0
+
+### Minor Changes
+
+- Adds new endpoints for on.auto
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/file-store@0.16.0
+  - @auto-engineer/message-bus@0.16.0
 
 ## 0.15.0