+++ packages/server-checks/CHANGELOG.md
@@ -1,12 +1,24 @@
 # @auto-engineer/server-checks
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @auto-engineer/cli@0.15.1
-  - @auto-engineer/message-bus@0.15.1
+  - @auto-engineer/message-bus@0.17.0
+  - @auto-engineer/cli@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/cli@0.16.0
+  - @auto-engineer/message-bus@0.16.0
 
 ## 0.15.0