+++ packages/server-generator-nestjs/CHANGELOG.md
@@ -1,11 +1,22 @@
 # @auto-engineer/server-generator-nestjs
 
-## 0.15.1
+## 0.17.0
+
+### Minor Changes
+
+- Adds new minimal example
 
 ### Patch Changes
 
 - Updated dependencies []:
-  - @auto-engineer/narrative@0.15.1
+  - @auto-engineer/narrative@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/narrative@0.16.0
 
 ## 0.15.0